functionalizing: moved orientation filter class into a single function. Moved exhaustion functions into util.energy. Refactored line end filter to work with orientation_filter function.
