+++ slam_recognition/filters/__init__.py
@@ -1,0 +1,1 @@
+from .orientation_filter import orientation_filter
