+++ slam_recognition/pyramid_filter.py
@@ -16,7 +16,7 @@ except ImportError:
 class PyramidFilter(object):
     callback_depth = 1
 
-    def __init__(self, output_size=(int(36*4), int(24*4)), output_colors=3, zoom_ratio=m.e ** .5):
+    def __init__(self, output_size=(int(36*8), int(24*8)), output_colors=3, zoom_ratio=m.e ** .5):
         """Generates several smaller images at different zoom levels from one input image."""
         self.output_size = output_size
         self.output_colors = output_colors
